рефакторинг для реализации через интерфейсы
